pyproject.toml
[project]
name = "lifeguard"
version = "0.1.0"
description = "Lightweight Intent-Focused Engine for Guidance in Unmanned Autonomous Rescue Deployments"
authors = [ { name = "John Seargeant" } ]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
dependencies = [
    "python-decouple",
    "vosk==0.3.45",
    "spacy==3.7.5",
    "pymavlink",
    "numpy<2.0",
    "scipy<1.14",
    "noisereduce",
    "pyaudio",
    "pynput",
    "pyttsx3",
    "word2number",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
